fixed signal handling to work properly in non-interactive mode

+++ src/signal_handlers.c
@@ -10,7 +10,6 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#define _GNU_SOURCE
 #include "minishell.h"
 #include <stddef.h>
 #include <signal.h>
@@ -19,25 +18,45 @@
 #include <readline/readline.h>
 #include <readline/history.h>
 
-volatile sig_atomic_t g_msh_signal = 0; 
+//volatile sig_atomic_t g_msh_signal = 0; 
 
-//
-void	sigint_handler_interactive(
+static void	sigint_handler_interactive(
 )
 {
-
 	write(STDOUT_FILENO, "\n", 1);
-	rl_replace_line("", 0); // Commented out - not available on all systems
+	rl_replace_line("", 0);
 	rl_on_new_line();
-	ioctl(STDIN_FILENO, TIOCSTI, "\n");
-//	g_msh_signal = SIGINT; 
+	rl_redisplay();
+}
+
+void	sigquit_handler_non_interactive(
+)
+{
+	if (g_msh_signal != SIGQUIT)
+	{
+		g_msh_signal = SIGQUIT;
+		kill(0, SIGQUIT);
+	}
+	else
+	{
+		g_msh_signal = 0;
+		write(STDOUT_FILENO, "\n", 1);
+	}
 }
 
 void	sigint_handler_non_interactive(
 )
 {
-	kill(0, SIGINT);
-	sigint_handler_interactive();
+	if (g_msh_signal != SIGINT)
+	{
+		g_msh_signal = SIGINT;
+		kill(0, SIGINT);
+	}
+	else
+	{
+		g_msh_signal = 0;
+		write(STDOUT_FILENO, "\n", 1);
+	}
 }
 
 void	handle_signals_child_process(
@@ -47,7 +66,6 @@ void	handle_signals_child_process(
 	struct sigaction	handle_sigint;
 
 	sigemptyset(&handle_sigint.sa_mask);
-	//handle_sigint.sa_handler = SIG_DFL;
 	handle_sigint.sa_handler = SIG_IGN;
 	handle_sigint.sa_flags = 0;
 	sigaction(SIGINT, &handle_sigint, NULL);
@@ -60,19 +78,23 @@ void	handle_signals_non_interactive(
 )
 {
 	struct sigaction	handle_sigquit;
+	struct sigaction	handle_sigint;
 
 	sigemptyset(&handle_sigquit.sa_mask);
-	//handle_sigquit.sa_handler = SIG_DFL;
-	handle_sigquit.sa_handler = SIG_IGN;
+	handle_sigquit.sa_handler = sigquit_handler_non_interactive;
 	handle_sigquit.sa_flags = 0;
 	sigaction(SIGQUIT, &handle_sigquit, NULL);
+
+	sigemptyset(&handle_sigint.sa_mask);
+	handle_sigint.sa_handler = sigint_handler_non_interactive;
+	handle_sigint.sa_flags = 0;
+	sigaction(SIGINT, &handle_sigint, NULL);
 }
 
 void	handle_signals_interactive(
 	void
 )
 {
-
 	struct sigaction	handle_sigint;
 	struct sigaction	handle_sigquit;
